requirements.txt
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
mysql-connector-python==8.0.32
networkx==3.0
nltk==3.8.1
numpy==1.24.2
Pillow==9.5.0
protobuf==3.20.3
pycodestyle==2.10.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
sympy==1.11.1
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
